refactor: simple refact.
