+++ Configurations/ActionTimelineReplacement.cs
@@ -1,0 +1,13 @@
+﻿namespace ActionTimelineReplacement.Configurations;
+
+public class ActionTimelineReplacement(
+    ushort animationStart,
+    ushort animationEnd,
+    ushort actionTimelineHit,
+    ushort castVfx)
+{
+    public ushort AnimationStart = animationStart;
+    public ushort AnimationEnd = animationEnd;
+    public ushort ActionTimelineHit = actionTimelineHit;
+    public ushort CastVfx = castVfx;
+}
